+++ doc/CHANGES.rst
@@ -14,6 +14,7 @@ Future
 
 * [FIX] update version if the version change
 * [REF] query.update and query.delete call the method on the Model
+* [REF] update apply state to do not raise on error in some case
 
 0.7.2 (2016-01-14)
 ------------------